Fix for exception when there are no sections
When there are no sections (due to no data) in the collectionview, the exception "request for number of items in section 0 when there are only 0 sections in the collection view" is thrown. This is typically thrown on line 95: var numberOfItemsInSection = collectionView?.numberOfItems(inSection: section) ?? 0

The fix is to check if numberOfSections does not equal zero via an 'if' statement. If it equals zero, it doesn't make sense to iterate through the entire function. Simply return an empty attributesArray instead.